backend/requirements.txt
fastapi
uvicorn
PyPDF2
sentence-transformers
qdrant-client
python-multipart
httpx
groq
google-generativeai
python-dotenv